renamed matrixlabs namespace to matrix
