+++ src/malos_base.h
@@ -12,9 +12,9 @@
 #include "./zmq_push.h"
 #include "./zmq_pull.h"
 
-#include "../protocol-buffers/matrixlabs/driver.pb.h"
+#include "../protocol-buffers/matrix/driver.pb.h"
 
-namespace pb = matrixlabs::malos::v1;
+namespace pb = matrix::malos::v1;
 
 namespace matrix_malos {
